tests/requirements.txt
conda-canary/label/dev::conda-build  # TEMPORARY; revert to release version after 24.3.0 is released
conda-forge::pytest-xprocess
conda-index
conda-lock
conda-forge::pre-commit
# needed for many conda tests
flask
